Move fallback renderer to its own folder in components
